data pull from private repo ++

+++ db/clone-data-repo.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
-# clone repo, if successful copy setup script to /docker-entrypoint-initdb.d/
+echo "checking if ssh key for the atlas content repo is present"
+FILE=/run/secrets/ssh_secret
+# if ssh key is not empty, init the db with data:
+if test -f "$FILE"; then
+    echo "ssh key present"
+    
+    # add ssh key and add github as known host
+    mkdir /root/.ssh/ && cp /run/secrets/ssh_secret /root/.ssh/id_rsa
+    chmod 400 /root/.ssh/id_rsa && ssh-keyscan github.com >> /root/.ssh/known_hosts
+
+    # clone repo, if successful copy setup script to /docker-entrypoint-initdb.d/
     git clone --single-branch --branch ${QC_ATLAS_CONTENT_REPOSITORY_BRANCH} ${QC_ATLAS_CONTENT_REPOSITORY_URL} ${QC_ATLAS_CONTENT_REPOSITORY_PATH}
 	if [ -d "${QC_ATLAS_CONTENT_REPOSITORY_PATH}/${QC_ATLAS_SUBFOLDER_CONTENT_BACKUP_FILES}" ]; then
         cp setup-atlas.sh /docker-entrypoint-initdb.d/
@@ -18,12 +28,15 @@
     else
 	    echo "unable to find specified directory with example data for qc-atlas in the repository"
 	fi
+else
+    echo "ssh key not present, proceeding with empty db for qc-atlas"
+fi
 
-# clone repo, if successful copy setup script to /docker-entrypoint-initdb.d/
-    git clone --single-branch --branch ${PATTERNATLAS_CONTENT_REPOSITORY_BRANCH} ${PATTERNATLAS_CONTENT_REPOSITORY_URL} ${PATTERNATLAS_CONTENT_REPOSITORY_PATH}
-    if [ -d "${PATTERNATLAS_CONTENT_REPOSITORY_PATH}/${PATTERNATLAS_SUBFOLDER_CONTENT_BACKUP_FILES}" ]; then
+ # clone repo, if successful copy setup script to /docker-entrypoint-initdb.d/
+git clone --single-branch --branch ${PATTERNATLAS_CONTENT_REPOSITORY_BRANCH} ${PATTERNATLAS_CONTENT_REPOSITORY_URL} ${PATTERNATLAS_CONTENT_REPOSITORY_PATH}
+if [ -d "${PATTERNATLAS_CONTENT_REPOSITORY_PATH}/${PATTERNATLAS_SUBFOLDER_CONTENT_BACKUP_FILES}" ]; then
     cp setup-patternatlas.sh /docker-entrypoint-initdb.d/
     echo "${PATTERNATLAS_CONTENT_REPOSITORY_BRANCH} of pattern-atlas-content repo was cloned successfully"
-    else
+else
     echo "unable to find specified directory with example data for pattern-atlas in the repository"
-    fi
+fi
